requirements.txt
# Flask web framework
flask==3.1.0

# Environment variables
python-dotenv==1.1.0

# Google Gemini AI
google-generativeai==0.8.5

# News & RSS feeds
requests==2.32.3
feedparser==6.0.11

# Google APIs (Drive + Calendar)
google-api-python-client==2.166.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Background scheduler
apscheduler==3.11.0
